fix(demo): give the docs their own root layout — kills the dark-shell theme collision
The landing's root layout forced bg-gray-950/text-gray-100 on every
route, so Fumadocs rendered light-theme text over the landing's dark
body: white sidebar glued to a black page, dark-on-dark paragraphs.

Route groups now carry two root layouts: (site) keeps the fixed dark
shell; (docs)/[lang] owns its <html> (lang per locale,
suppressHydrationWarning, Inter) and lets RootProvider drive real
light/dark theming. Verified by computed styles + screenshots in both
schemes: light = dark text on rgb(245,245,245), dark = light text on
rgb(18,18,18), landing untouched, sidebar at Fumadocs' standard 268px
with clean collapse/expand.
