+++ apps/demo/src/app/(docs)/[lang]/docs.css
@@ -1,0 +1,6 @@
+@import "fumadocs-ui/style.css";
+@import "fumadocs-ui/css/neutral.css";
+
+body {
+  font-family: var(--font-inter), ui-sans-serif, system-ui, -apple-system, sans-serif;
+}
