+++ apps/demo/src/app/(docs)/[lang]/docs/[[...slug]]/page.tsx
