+++ apps/demo/src/app/(docs)/[lang]/docs/docs-widget.tsx
