+++ apps/demo/src/app/(site)/demo/diagnostics-triggers.tsx
