--- apps/demo/src/app/[lang]/docs/docs.css
@@ -1,2 +1,0 @@
-@import "fumadocs-ui/style.css";
-@import "fumadocs-ui/css/neutral.css";
